Move MenuOptions out of the models folder
